remove Suites and Tags
